Removed quadrotor task and renamed pendulum
