+++ leap_c/examples/__init__.py
@@ -1,6 +1,5 @@
 import leap_c.examples.mujoco.task
 import leap_c.examples.pointmass.task
-import leap_c.examples.pendulum_on_a_cart.task
-import leap_c.examples.quadrotor.task
+import leap_c.examples.cartpole.task
 import leap_c.examples.chain.task
 import leap_c.examples.mujoco.reacher.task
