--- leap_c/examples/quadrotor/init_iterate.json
@@ -1,293 +1,0 @@
-{
-    "lam_0": [
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        2.2133134127256633e-19,
-        0.684656018815078,
-        1e-16,
-        1.5527586083844284e-18,
-        1.1296627903201004e-18,
-        1e-16,
-        1e-16,
-        4.4900651568929875e-20,
-        0.22628167848192574,
-        1.560501695896745e-19,
-        1.1161326198530946e-19,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        8.501212511192925e-20,
-        1e-16,
-        1e-16,
-        6.006264734487435e-43,
-        1e-16,
-        1e-16,
-        4.547170609969349e-22,
-        1.660517384493958e-20,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        8.853904104951952e-20
-    ],
-    "lam_1": [
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        8.458647171731975,
-        8.458647288823709
-    ],
-    "lam_2": [
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        8.44679775510849,
-        8.446797874446501
-    ],
-    "lam_3": [
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        8.434886692856098,
-        8.434886814455986
-    ],
-    "lam_4": [
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        1e-16,
-        8.422914126301249,
-        8.422914246874873
-    ],
-    "lam_5": [
-        1e-16,
-        1e-16
-    ],
-    "pi_0": [
-        -8.501212511192925e-20,
-        2.2133134127256633e-19,
-        0.6789294409642698,
-        -6.006264734487435e-43,
-        1.584522619494016e-18,
-        1.141360179145596e-18,
-        -4.547170609969349e-22,
-        -1.3204688840462412e-20,
-        3.604739791802722e-20,
-        0.19864244223919178,
-        1.2466577725639398e-19,
-        8.889858445675801e-20,
-        -8.852994670829958e-20
-    ],
-    "pi_1": [
-        -8.501946078595864e-20,
-        2.213295912984696e-19,
-        0.6730776195767689,
-        -1.4654210642011616e-42,
-        1.605686507582074e-18,
-        1.1441690151771418e-18,
-        -4.543312028152004e-22,
-        -9.805804200661293e-21,
-        2.719262864078974e-20,
-        0.17128402889823943,
-        9.242561094991845e-20,
-        6.54863286383654e-20,
-        -8.813232406300475e-20
-    ],
-    "pi_2": [
-        -8.516542589443028e-20,
-        2.2137861076212326e-19,
-        0.6671071039045959,
-        -2.6098341485924614e-42,
-        1.648322071274927e-18,
-        1.1691561980530357e-18,
-        -4.531716336024764e-22,
-        -6.400737856931344e-21,
-        1.8335095166948082e-20,
-        0.1441446377216225,
-        6.285795830379267e-20,
-        4.5284762847216e-20,
-        -8.77338587832146e-20
-    ],
-    "pi_3": [
-        -8.494717699891552e-20,
-        2.207821724016396e-19,
-        0.6609981265588553,
-        -5.1350337274722805e-42,
-        1.6760681356434321e-18,
-        1.1814842492946062e-18,
-        -4.508897697863246e-22,
-        -2.9842256272828033e-21,
-        9.47553200787264e-21,
-        0.11716443445804964,
-        2.5424855253532613e-20,
-        1.7751071478576142e-20,
-        -8.697443867147656e-20
-    ],
-    "pi_4": [
-        -8.413768521632952e-20,
-        2.1911599541855446e-19,
-        0.6547053011933148,
-        6.751035518611872e-45,
-        1.6598995658766964e-18,
-        1.1611620146075643e-18,
-        -4.475687735922738e-22,
-        3.886263628363357e-22,
-        6.899024310049222e-22,
-        0.09028640886323785,
-        -7.46732352674306e-21,
-        -4.458653952905379e-21,
-        -8.664983016534499e-20
-    ],
-    "u_0": [
-        970.4363135003517,
-        970.4363135003517,
-        970.4363135003517,
-        970.4363135003517
-    ],
-    "u_1": [
-        970.4363963422223,
-        970.4363963422223,
-        970.4363963422223,
-        970.4363963422223
-    ],
-    "u_2": [
-        970.436480212674,
-        970.436480212674,
-        970.436480212674,
-        970.436480212674
-    ],
-    "u_3": [
-        970.4365653176078,
-        970.4365653176078,
-        970.4365653176078,
-        970.4365653176078
-    ],
-    "u_4": [
-        970.4366518631784,
-        970.4366518631784,
-        970.4366518631784,
-        970.4366518631784
-    ],
-    "x_0": [
-        0.0,
-        0.0,
-        5.726577850808211e-07,
-        1.0,
-        0.0,
-        0.0,
-        0.0,
-        0.0,
-        0.0,
-        4.820586041631712e-07,
-        0.0,
-        0.0,
-        0.0
-    ],
-    "x_1": [
-        1.6392529610492625e-27,
-        -9.521551126017612e-28,
-        5.910930694445207e-07,
-        1.0,
-        3.726957569498566e-25,
-        6.323004145272289e-25,
-        -3.9290006245437714e-25,
-        1.6655693608474355e-25,
-        -9.290888197330989e-26,
-        4.3970561402180425e-07,
-        3.7113267639781546e-23,
-        6.241336215654017e-23,
-        -3.929027130270187e-23
-    ],
-    "x_2": [
-        1.497170593072539e-26,
-        -4.8511834186713664e-27,
-        6.091741324531084e-07,
-        1.0,
-        -2.5340262051875028e-24,
-        -1.892539158320177e-24,
-        -1.1820428118996063e-24,
-        1.624543959200065e-25,
-        2.8083028362946404e-25,
-        4.643475364075836e-07,
-        -3.3009030084840704e-22,
-        -3.182577406797774e-22,
-        -3.968508040854688e-23
-    ],
-    "x_3": [
-        -2.248975193886999e-26,
-        6.141533515734388e-26,
-        6.295974071642521e-07,
-        1.0,
-        -1.734898306839475e-24,
-        -8.902435417115887e-25,
-        -2.3507370441535647e-24,
-        -1.9245406021946235e-24,
-        2.9084295032795136e-24,
-        5.568161991496005e-07,
-        4.1352204772010687e-22,
-        4.0564969551343155e-22,
-        -7.729421970246311e-23
-    ],
-    "x_4": [
-        -8.426667212892589e-26,
-        1.7344657123844312e-25,
-        6.550959300299815e-07,
-        1.0,
-        2.1004585960205572e-24,
-        2.222595279403282e-24,
-        -3.457458864620035e-24,
-        -7.652877304764541e-25,
-        2.1847284768241326e-24,
-        7.181099441368736e-07,
-        -2.8562570785263054e-23,
-        -1.0508424953691238e-22,
-        -3.2883193517594745e-23
-    ],
-    "x_5": [
-        -8.847378079841823e-26,
-        2.3040829443125647e-25,
-        6.884460055729915e-07,
-        1.0,
-        1.7454436732027227e-24,
-        1.221003326722151e-24,
-        -4.706345493723486e-24,
-        4.086544994631638e-25,
-        7.254570445584262e-25,
-        9.49393833013624e-07,
-        -7.852157367501663e-24,
-        -4.6884338625564186e-24,
-        -9.111539092807181e-23
-    ]
-}
