--- leap_c/examples/quadrotor/model_params.yaml
@@ -1,48 +1,0 @@
-bound_z: 10.0
-cx1_fres: -0.15019502977978097
-cx2_fres: -0.002709505498055363
-cx3_fres: -7.016151805e-08
-cy1_fres: -0.16901964908034103
-cy2_fres: 0.000304255916122357
-cy3_fres: -7.6767209802e-08
-cz1_fres: 0.08750617566879396
-cz2_fres: -0.568767683352191
-cz3_fres: -9.98613822031e-07
-cz4_fres: 1.32578769797e-07
-cz5_fres: 0.006458575813864983
-cz6_fres: -1.2099005181e-08
-cz7_fres: -9.475727292728e-05
-gravity: 9.81
-inertia:
-- - 0.002284
-  - 0
-  - 0
-- - 0
-  - 0.00165
-  - 0
-- - 0
-  - 0
-  - 0.003688
-inertia_motor: 5.64e-06
-mass: 0.6
-motor_cl: 1.562522e-06
-motor_kappa:
-- 0.01125
-- -0.01125
-- 0.01125
-- -0.01125
-motor_omega_max: 2332
-rotor_loc:
-- - 0.075
-  - 0.1
-  - 0
-- - 0.075
-  - -0.1
-  - 0
-- - -0.075
-  - -0.1
-  - 0
-- - -0.075
-  - 0.1
-  - 0
-t_motor: 0.022
