+++ tests/leap_c/acados/test_layer.py
@@ -1,7 +1,7 @@
 import numpy as np
 import torch
 from leap_c.examples.chain.mpc import ChainMpc
-from leap_c.examples.pendulum_on_a_cart.mpc import PendulumOnCartMPC
+from leap_c.examples.cartpole.mpc import PendulumOnCartMPC
 from leap_c.examples.pointmass.mpc import PointMassMPC
 from leap_c.ocp.acados.mpc import MpcInput, MpcParameter
 from leap_c.ocp.acados.layer import MpcSolutionModule
